Move AMI related types into their own file
Also require instances to specify amis, rather than defaulting
to legacy choices
