+++ library/aws/autoscale-deployment.ts
@@ -12,10 +12,10 @@ import * as policies from './policies.ts';
 import * as docker from '../docker.ts';
 import * as camus2 from '../camus2/camus2.ts';
 import * as C from '../../library/camus2/adl-gen/config.ts';
+import * as amis from './amis.ts';
 
 import {
   EndPoint,
-  getDefaultAmi,
   httpsFqdnsFromEndpoints,
   ec2InstallScript,
 } from './ec2-deployment.ts';
@@ -179,8 +179,7 @@ export function createController(
 
   const controller = aws.createInstanceWithEip(tfgen, controller_label, sr, shared.externalSubnetIds(sr)[0], {
     instance_type: AT.t2_micro,
-    ubuntu_version: params.controller_ubuntu_version, 
-    ami: params.controller_amis || getDefaultAmi,
+    ami: params.controller_amis,
     security_group: sr.bastion_security_group,
     key_name: params.key_name,
     customize_instance: (i: AR.InstanceParams) => {
@@ -271,9 +270,7 @@ export function createProcessorAutoScaleGroup(
   const launch_config_params = {
     name_prefix: tfgen.scopedName(name).join('-') + '-',
     key_name: params.key_name,
-    image_id: params.appserver_amis
-      ? params.appserver_amis(sr.region)
-      : getDefaultAmi(sr.region, params.appserver_ubuntu_version),
+    image_id: params.appserver_amis(sr.region),
     instance_type: params.appserver_instance_type,
     iam_instance_profile: instance_profile.id,
     security_groups: [sr.appserver_security_group.id],
@@ -692,12 +689,11 @@ export interface AutoscaleProcessorParams {
    */
   controller_extra_policies?: policies.NamedPolicy[];
 
-  controller_ubuntu_version: aws.UbuntuVersion;
   /**
    * Specifies the AMI for the controller. Defaults to an ubuntu 16.04 AMI
    * for the appropriate region.
    */
-  controller_amis?(region: AT.Region): AT.Ami;
+  controller_amis: amis.AmiSelector;
 
   /**
    * The context files are fetched from S3 and made available to the controller instance for
@@ -720,12 +716,11 @@ export interface AutoscaleProcessorParams {
    */
   appserver_instance_type: AT.InstanceType;
 
-  appserver_ubuntu_version: aws.UbuntuVersion;
   /**
    * Specifies the AMI for the EC2 instances. Defaults to an ubuntu 16.04 AMI
    * for the appropriate region.
    */
-  appserver_amis?(region: AT.Region): AT.Ami;
+  appserver_amis: amis.AmiSelector;
 
   /**
    * The context files are fetched from S3 and made available to hx-deploy-tool for interpolation
