+++ library/aws/ec2-deployment.ts
@@ -13,6 +13,7 @@ import * as bootscript from '../bootscript.ts';
 import * as docker from '../docker.ts';
 import * as camus2 from '../camus2/camus2.ts';
 import * as C from '../../library/camus2/adl-gen/config.ts';
+import * as amis from './amis.ts';
 
 /**
  *  Creates a logical deployment on a single publicly addressable EC2 instance, including:
@@ -37,8 +38,7 @@ export function createEc2Deployment(
   const instance_profile = createIamInstanceProfile(tfgen, name, sr, params);
   const appserver = aws.createInstanceWithEip(tfgen, name, sr, params.subnet_id, {
     instance_type: params.instance_type,
-    ubuntu_version: params.ubuntu_version,
-    ami: params.ami || getDefaultAmi,
+    ami: params.ami,
     security_group: sr.appserver_security_group,
     key_name: params.key_name,
     customize_instance: (i: AR.InstanceParams) => {
@@ -63,8 +63,7 @@ export function createInternalEc2Deployment(
   const instance_profile = createIamInstanceProfile(tfgen, name, sr, params);
   const appserver = aws.createInstance(tfgen, name, sr, params.subnet_id, {
     instance_type: params.instance_type,
-    ubuntu_version: params.ubuntu_version,
-    ami: params.ami || getDefaultAmi,
+    ami: params.ami,
     security_group: sr.appserver_security_group,
     key_name: params.key_name,
     customize_instance: (i: AR.InstanceParams) => {
@@ -277,8 +276,6 @@ export interface Ec2InstanceDeploymentParams {
   // The AWS keyname used for the EC2 instance.
   key_name: AT.KeyName;
 
-  ubuntu_version: aws.UbuntuVersion;
-  // The AWS instance type (ie mem and cores) for the EC2 instance.
   instance_type: AT.InstanceType;
   // The subnet to be used
   subnet_id: AR.SubnetId,
@@ -293,9 +290,8 @@ export interface Ec2InstanceDeploymentParams {
   //    main:   ${dns_name}.${primary_dns_zone}
   //    test:   ${dns_name}-test.${primary_dns_zone}
   endpoints: EndPoint[];
-  // Specifies the AMI for the EC2 instance. Defaults to an ubuntu 16.04 AMI
-  // for the appropriate region.
-  ami?(region: AT.Region): AT.Ami;
+  // Specifies the AMI for the EC2 instance. 
+  ami: amis.AmiSelector;
   // The EC2 instance created is given an IAM profile with sufficient access policies to
   // log metrics, run the deploy tool and create SSL certificates. Additional policies
   // can be specified here.
@@ -394,112 +390,6 @@ export function endPointsSummary(eps: EndPoint[]) : {[key:string] : string[]} {
 export interface Ec2Deployment {
   eip: AR.Eip;
   ec2: AR.Instance;
-}
-
-
-/** created from https://cloud-images.ubuntu.com/locator/ec2/
- * filter version and arch, copy-paste and a bit of reformating
- */
-const ec2_ami_1604_amd64 = [
-  { region: "af-south-1", ami: "ami-0805fe821528cb0ff"},
-  { region: "ap-east-1", ami: "ami-01e6a2bfeab33d1c4"},
-  { region: "ap-northeast-1", ami: "ami-0e42827f7b2eaa246"},
-  { region: "ap-south-1", ami: "ami-01220a0a70c164303"},
-  { region: "ap-southeast-1", ami: "ami-0c21f9b6eb55124d4"},
-  { region: "ca-central-1", ami: "ami-0baa2760c1decf0c8"},
-  { region: "eu-central-1", ami: "ami-01299931803ce83f6"},
-  { region: "eu-north-1", ami: "ami-0567085e558e02053"},
-  { region: "eu-south-1", ami: "ami-0f5cea12ee799691a"},
-  { region: "eu-west-1", ami: "ami-016ee74f2cf016914"},
-  { region: "me-south-1", ami: "ami-063e418f8a67c299b"},
-  { region: "sa-east-1", ami: "ami-0c9bb0eaa91f06fbc"},
-  { region: "us-east-1", ami: "ami-0133407e358cc1af0"},
-  { region: "us-west-1", ami: "ami-0fdf8b5989f22a4e0"},
-  { region: "cn-north-1", ami: "ami-03bd5b54f08201029"},
-  { region: "cn-northwest-1", ami: "ami-01924473944c652d8"},
-  { region: "us-gov-east-1", ami: "ami-0fe6a74784a689a73"},
-  { region: "us-gov-west-1", ami: "ami-086b498deca4cc63c"},
-  { region: "ap-northeast-2", ami: "ami-0daccca4e1fc56e3f"},
-  { region: "ap-southeast-2", ami: "ami-0e554a91eb4e7b6d7"},
-  { region: "eu-west-2", ami: "ami-0ad9f4c7544ed8cea"},
-  { region: "us-east-2", ami: "ami-01685d240b8fbbfeb"},
-  { region: "us-west-2", ami: "ami-079e7a3f57cc8e0d0"},
-  { region: "ap-northeast-3", ami: "ami-008ddf50457c78d08"},
-  { region: "eu-west-3", ami: "ami-062a3f6e040d4c62a"},
-]
-
-const ec2_ami_2004_amd64 = [
-  {region: "af-south-1", ami: "ami-0982b51b05c9be169"},
-  {region: "ap-east-1", ami: "ami-04c4bc345657bf245"},
-  {region: "ap-northeast-1", ami: "ami-0b0ccc06abc611fa0"},
-  {region: "ap-south-1", ami: "ami-0443fb07ed652c341"},
-  {region: "ap-southeast-1", ami: "ami-0f0b17182b1d50c14"},
-  {region: "ca-central-1", ami: "ami-04673916e7c7aa985"},
-  {region: "eu-central-1", ami: "ami-05e1e66d082e56118"},
-  {region: "eu-north-1", ami: "ami-00888f2a5f9be4390"},
-  {region: "eu-south-1", ami: "ami-06a3346e9e869f9b1"},
-  {region: "eu-west-1", ami: "ami-0298c9e0d2c86b0ed"},
-  {region: "me-south-1", ami: "ami-0420827ce9e4a7552"},
-  {region: "sa-east-1", ami: "ami-04e56ee48b28650b3"},
-  {region: "us-east-1", ami: "ami-019212a8baeffb0fa"},
-  {region: "us-west-1", ami: "ami-0b08e71a81ba4200f"},
-  {region: "cn-north-1", ami: "ami-0741e7b8b4fb0001c"},
-  {region: "cn-northwest-1", ami: "ami-0883e8062ff31f727"},
-  {region: "us-gov-east-1", ami: "ami-0fe6338c47e61cd5d"},
-  {region: "us-gov-west-1", ami: "ami-087ee83c8de303181"},
-  {region: "ap-northeast-2", ami: "ami-0f49ee52a88cc2435"},
-  {region: "ap-southeast-2", ami: "ami-04b1878ebf78f7370"},
-  {region: "eu-west-2", ami: "ami-0230a6736b38ae83e"},
-  {region: "us-east-2", ami: "ami-0117d177e96a8481c"},
-  {region: "us-west-2", ami: "ami-02868af3c3df4b3aa"},
-  {region: "ap-northeast-3", ami: "ami-01ae085ceefba2dbf"},
-  {region: "eu-west-3", ami: "ami-06d3fffafe8d48b35"},
-]
-
-export type UbuntuVersion = "1604" | "2004";
-
-/**
- * Standard ubuntu base AMIs.
- *
- * (ubuntu xenial, hvm:ebs-ssd, EBS General purpose SSD, x86)
- *  https://cloud-images.ubuntu.com/locator/ec2/
- */
-export function getDefaultAmi(region: AT.Region, ubuntu_version: UbuntuVersion
-  //  = "1604"
-   ): AT.Ami {
-  if( ubuntu_version === "1604" ) {
-    if (region.value === AT.ap_southeast_2.value) {
-      return AT.ami('ami-47c21a25');
-    }
-    if (region.value === AT.us_east_1.value) {
-      return AT.ami('ami-03a935aafa6b52b97');
-    }
-    if (region.value === AT.us_east_2.value) {
-      return AT.ami('ami-5e8bb23b');
-    }
-    if (region.value === AT.ca_central_1.value) {
-      return AT.ami('ami-01957f6afe4e49edd');
-    }
-    if (region.value === AT.eu_west_2.value) {
-      return AT.ami('ami-0fab23d0250b9a47e');
-    }
-    if (region.value === AT.eu_west_1.value) {
-      return AT.ami('ami-0f2ed58082cb08a4d');
-    }
-    if (region.value === AT.eu_north_1.value) {
-      return AT.ami('ami-04b331702444679c3');
-    }
-    if (region.value === AT.eu_central_1.value) {
-      return AT.ami('ami-05ed2c1359acd8af6');
-    }
-  }
-  const ec2_ami_data = ubuntu_version === "1604" ? ec2_ami_1604_amd64 : ec2_ami_2004_amd64;
-  for (const ec2_ami of ec2_ami_data) {
-    if(region.value === ec2_ami.region ) {
-      return AT.ami(ec2_ami.ami);
-    }
-  }
-  throw new Error('No AMI specified for region ' + region.value);
 }
 
 export function ec2InstallScript(
